aigateway list model add user preference
